requirements.txt
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
colour-science>=0.4.3
flask>=2.0.0
flask-cors>=4.0.0